sort out blockdev_direct_IO variants
Move the call to vmtruncate to get rid of accessive blocks to the
callers
in prepearation of the new truncate calling sequence.  This was only
done
for DIO_LOCKING filesystems, so the __blockdev_direct_IO_newtrunc
variant
was not needed anyway.  Get rid of blockdev_direct_IO_no_locking and
its _newtrunc variant while at it as just opencoding the two additional
paramters is shorted than the name suffix.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>