remove inode_setattr
Replace inode_setattr with opencoded variants of it in all callers.
This
moves the remaining call to vmtruncate into the filesystem methods where
it
can be replaced with the proper truncate sequence.

In a few cases it was obvious that we would never end up calling
vmtruncate
so it was left out in the opencoded variant:

spufs: explicitly checks for ATTR_SIZE earlier
btrfs,hugetlbfs,logfs,dlmfs: explicitly clears ATTR_SIZE earlier
ufs: contains an opencoded simple_seattr + truncate that sets the
filesize just above

In addition to that ncpfs called inode_setattr with handcrafted iattrs,
which allowed to trim down the opencoded variant.

Signed-off-by: Christoph Hellwig <hch@lst.de>