Audit: do not get and put just to free a watch
deleting audit watch rules is not currently done under
audit_filter_mutex.
It was done this way because we could not hold the mutex during inotify
manipulation.  Since we are using fsnotify we don't need to do the extra
get/put pair nor do we need the private list on which to store the
parents
while they are about to be freed.

Signed-off-by: Eric Paris <eparis@redhat.com>