xfs: new truncate sequence
Convert XFS to the new truncate sequence.  We still can have errors
after
updating the file size in xfs_setattr, but these are real I/O errors and
lead
to a transaction abort and filesystem shutdown, so they are not an
issue.

Errors from ->write_begin and write_end can now be handled correctly
because
we can actually get rid of the delalloc extents while previous the
buffer
state was stipped in block_invalidatepage.

There is still no error handling for ->direct_IO, because doing so will
need
some major restructuring given that we only have the iolock shared and
do not
hold i_mutex at all.  Fortunately leaving the normally allocated blocks
behind
there is not a major issue and this will get cleaned up by
xfs_free_eofblock
later.

Note: the patch is against Al's vfs.git tree as that contains the
nessecary
preparations.  I'd prefer to get it applied there so that we can get
some
testing in linux-next.

Signed-off-by: Christoph Hellwig <hch@lst.de>