cifs: define superblock-level cache index objects and register them
Define superblock-level cache index objects (managed by cifsTconInfo
structs).
Each superblock object is created in a server-level index object and in
itself
an index into which inode-level objects are inserted.

The superblock object is keyed by sharename. The UniqueId/IndexNumber is
used to
validate that the exported share is the same since we accessed it last
time.

Signed-off-by: Suresh Jayaraman <sjayaraman@suse.de>
Signed-off-by: Steve French <sfrench@us.ibm.com>